pyproject.toml
[project]
name = "x-twitter-api-v2-demo"
version = "0.1.0"
description = "A demo illustrating how to tweet via X's new Twitter v2 API using Python Flask"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests-oauthlib>=1.3.1",
    "requests>=2.32.3",
    "tweepy>=4.14.0",
]
